+++ packages/tools/utils/src/modules/[any-number]-[array-number].js
@@ -5,14 +5,42 @@
 import * as _ from 'lamb';
 
 import {getLength} from './iterable-number';
-import {arraySum} from './array-number';
+import {isNumber} from './any-boolean';
+
+/**
+ * Return a function expecting an array and summing the numbers obtained
+ * from applying the provided accessor to the array items.
+ * Note that it skips items where the accessor does not return a number.
+ *
+ * @function
+ * @arg {function} accessor - Any -> Number
+ * @return {function} - Array -> Number
+ *
+ * @example
+> sumValues = arraySumWith(_.getKey('a'))
+> sumValues([{a: 1}, {a: 2}, {a: 3}])
+6
+> sumValues([{a: 1}, {a: 2}, {a: 'hey'}])
+3
+> sumValues([{a: 1}, {a: 2}, {notA: 3}])
+3
+> sumValues([{a: 'hey'}, {notA: 'b'}, {notA: 3}])
+0
+ *
+ * @since 0.16.0
+ */
+export const arraySumWith = accessor => _.reduceWith((acc, item) => {
+	const value = accessor(item);
+
+	return acc + (isNumber(value) ? value : 0);
+}, 0)
 
 /**
  * Return the average of values of a {key, value}[] array
  *
  * @function
- * @arg {array} – {key, value}[]
- * @return {number}
+ * @arg {function} accessor - Any -> Number
+ * @return {function} - Array -> Number
  *
  * @example
 > makeAverageOfA = makeAverageWith(_.getKey('a'));
@@ -30,10 +58,7 @@ import {arraySum} from './array-number';
  *
  * @since 0.11.0
  */
-export const makeAverageWith = getter => _.pipe([
-	_.collect([
-		_.pipe([_.mapWith(getter), arraySum]),
-		getLength
-	]),
+export const makeAverageWith = accessor => _.pipe([
+	_.collect([arraySumWith(accessor), getLength]),
 	_.apply(_.divide),
 ]);